pyproject.toml
[project]
name = "tttsvm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "httpx>=0.28.1",
    "numpy>=2.3.2",
    "openai>=1.99.9",
    "ormsgpack>=1.10.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pynput>=1.8.1",
    "pyperclip>=1.9.0",
    "pystray>=0.19.5",
    "pyttsx3>=2.99",
    "requests>=2.32.4",
    "sounddevice>=0.5.2",
    "websockets>=15.0.1",
]
